Класс OsmParser.java: * Появилась возможность задавать конфигурацию, для установки максимальной скорости на дороге.
Класс ParseHelper.java переименован в OsmParseHelper.java. По причине появления еще одного "помощника" для парсинга файлов (ConfigFileParse.java)

Класс MaxSpeedConfiguration.java. Класс создан для предоставления информации о конфигурировании, выдачи скорости по дороге, при парсинге дороги.

Класс BlockedPointConfiguration.java. Не реализованный класс, требуется дописать. Задумывается как для предоставления информации о конфигурировании, информация о блокированных точках дорог.

Под новый функционал написан тест и созданы тестовые данные.

+++ src/main/java/mapObjects/Way.java
@@ -13,7 +13,7 @@ public class Way {
     //private byte direction; //https://wiki.openstreetmap.org/wiki/RU:Key:oneway?uselang=ru
     private final boolean oneWay;
 
-    public Way(long id, List<Long> idNodes, int wayType, int speed, boolean oneWay){
+    public Way(long id, List<Long> idNodes, int wayType, double speed, boolean oneWay){
         this.id = id;
         this.idNodes = idNodes;
         this.wayType = wayType;
